requirements.txt
Django==2.1
Markdown==2.6.8
Pillow == 4.0.0
argparse==1.2.1
csscompressor==0.9.4
django-autoslug==1.9.3
django-markitup==3.0.0
django-pipeline==1.6.11
-e git+https://github.com/mysociety/django-pipeline-csscompressor#egg=django_pipeline_csscompressor-master
futures==3.0.5
libsass==0.12.3
olefile==0.44
six==1.10.0
django-import-export==1.2.0
git+https://github.com/ajparsons/useful_inkleby@d2#egg=useful_inkleby
beautifulsoup4
django-bootstrap-form
django-paypal
django-ipware==2.1.0
cobble==0.1.3
parsimonious==0.8.1
jsmin==2.2.2
dirsync==2.2.3
html2text
selenium==3.13
tinify==1.5.1
invoke
unicodecsv
anyjson
django-htmlmin
markdownify
git+https://github.com/gaojiuli/tomd#egg=tomd
django-bootstrap-form
ruamel.yaml
PyPDF2==1.26.0
PyMuPDF==1.14.13
mammoth==1.4.9